Use BigDecimal instead of double for price handling, remove unnecessary imports

+++ interfaces/src/main/java/ar/edu/itba/paw/persistance/ProductDao.java
@@ -2,11 +2,12 @@ package ar.edu.itba.paw.persistance;
 
 import ar.edu.itba.paw.model.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface ProductDao {
-    Product create(int categoryId, String name, String description, int imageId, double price);
+    Product create(int categoryId, String name, String description, int imageId, BigDecimal price);
 
     Optional<Product> getById(int productId);
 
@@ -14,7 +15,7 @@ public interface ProductDao {
 
     List<Product> getByRestaurantOrderByCategoryOrder(int restaurantId);
 
-    boolean updatePrice(int productId, double price);
+    boolean updatePrice(int productId, BigDecimal price);
 
     boolean updateName(int productId, String name);
